SQL change: QCheckBoxes cannot contain NULL
if a field that is mapped to a QCheckBox contains "NULL", it will not change to "unchecked". It will simply not change, meaning that if the previous selected row was "checked", the current row will still show "checked". If the previous selected row was "unchecked", than the current row will continue to show "unchecked".

So "NULL" in a column that is mapped to a QCheckBox will result in strange behaviour in the program.