Added property to BTListView to link the model/view to the database
the database contains a list of IDs that should be selected in the view.

At the moment, it does not work correct yet: the db can contain several IDs, at the moment only 1 ID gets selected instead of all
problem lies with `this->setCurrentIndex(this->model()->index( i, 1));` at line 97 in btlistview.cpp ?

Maybe we can check the items instead of selecting them? Could be more userfriendly as well.