When clicking on SAVE for the second time, it has no response #2
[PROBLEM]



Choose a user, change some information and press the SAVE button.

The information gets saved to the database.

Now again, change some information and press the SAVE button.

The change is visible on the screen, but does not get saved to the database.


[SUGGESTED SOLUTION]

When clicking the SAVE button, the model tries to load an invalid QModelIndex (-1, -1). I do not know why.

Now, using a new class member variable `last_known_index` to keep track which row you were changing.

When the SAVE button is clicked, this member variable makes sure that the line with that index is loaded, so when additional changes are being made, the program will now know to which row they apply, and they will get saved to the database.