array of booleans infowindows[]
issue #5 : array of booleans created, when a marker is clicked to open an infowindow, the value is set to TRUE.

But now, it stays TRUE forever. What if the user closes the infowindow and later wants to reopen it? It won't work.

Maybe better to reset the boolean to FALSE when the marker is clicked again, and then also CLOSE that infowindow.
-> problem: the infowindow is only declared locally... Maybe an array of infowindows???

Can we test "if (infowindow.open() ) then infowindow.close() "??



    for (var i = 0; i < deelnemers.length; i++) {
 			  var deelnemer = deelnemers[i];
                          var infowindow = new google.maps.InfoWindow({ content: deelnemers[i][3] });
                          infowindows.push([infowindow, false]);
                          ...

and in return `function() {`:


                                      if (!infowindows[i][2])
                                      {
                                          infowindow[i][1].open(map, extramarker);
                                          infowindows[i][2] = true;
                                      }